// src/app/shared/interfaces/horaire.interface.ts
export interface HoraireInterface {
    _id: string;
    oLundi: string;
    fLundi: string;
    oMardi: string;
    fMardi: string;
    oMercredi: string;
    fMercredi: string;
    oJeudi: string;
    fJeudi: string;
    oVendredi: string;
    fVendredi: string;
    oSamedi: string;
    fSamedi: string;
    oDimanche: string;
    fDimanche: string;
  }